python_server/pyproject.toml
[project]
name = "gameserver"
version = "0.1.0"
description = "Multiplayer Tower-Defense Empire-Building Game Server"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "websockets>=12.0",
    "aiosqlite>=0.19",
    "pyyaml>=6.0",
    "msgpack>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.scripts]
gameserver = "gameserver.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.mypy]
python_version = "3.9"
strict = true
